Made Android not force crash if a test loop is active

+++ platform/android/java/lib/src/org/godotengine/godot/utils/ProcessPhoenix.java
@@ -106,9 +106,13 @@ public final class ProcessPhoenix extends Activity {
    * @param pid
    */
   public static void forceQuit(Activity activity, int pid) {
-    Process.killProcess(pid); // Kill original main process
-    activity.finishAndRemoveTask();
-    Runtime.getRuntime().exit(0); // Kill kill kill!
+  	if (activity.getIntent().getAction().equals("com.google.intent.action.TEST_LOOP")) {
+	    Process.killProcess(pid); // Kill original main process
+	    activity.finishAndRemoveTask();
+	    Runtime.getRuntime().exit(0); // Kill kill kill!
+	    return;
+	}
+	throw new RuntimeException("ForceQuit Called, throwing an exception rather than a graceful exit so Crashlytics can capture it");
   }
 
   // -- GODOT end --
